Fixed issue where 5 starrs was always saved to the db

+++ app/src/main/java/com/techflix/group36/techflix/Movie.java
@@ -39,7 +39,7 @@ public class Movie implements Serializable {
         return Rating.filterRatingsByMovie(this);
     }
 
-    public void rateMovie(int stars, String comment) {
+    public void rateMovie(float stars, String comment) {
         Rating newRating = new Rating(stars, comment, User.currentUser, this);
         newRating.save();
     }
